Use Azure AD for login
We want to use Azure AD SSO for the pre-AB services (Transfers and Conversions). Initially, we have decided to add it separately to the two apps, using the same ClientId, as it will be the same users using both apps.

Add the Microsoft.Identity.Web package for integration with Microsoft identity platform (Azure AD), and Microsoft.Identity.Web.UI which add an AccountController and associated views.

Override the `SignedOut` page to provide a custom view for when the user is redirected following sign out.

Remove the previous implementation of the login page and any references.