Remove appsettings.json to fix integration tests
The ApplyToBecomeInternal.Tests project includes its own appsettings.json which overrides the one in ApplyToBecomeInternal.

However, the integration tests use the StartUp of the ApplyToBecomeInternal project, which will return an argument null exception when the "AzureAd" section of the appsettings.json is not included, and is currently causing all of the tests to break.

The "TramsApi:Endpoint" configuration is overridden separately in the integration tests setup anyway via AddInMemoryCollection, so it should be okay to remove appsettings.json entirely from the test project.