Fix production unhandled error page
We use two pieces of middleware to handle errors. `UseExceptionHandler` for unhandled errors, and `UseStatusCodePagesWithReExecute` when the response has a status code.

UseExceptionHandler was still pointing to the error page generated by  the project template, meaning that users would see a non-custom page when the error is unhandled.

Instead, point UseExceptionHandler to the same page that we use for our error code pages, and apply a default title of "An error occurred while processing your request". Remove the template error page as it isn't used anymore.