Show "Page not found" when redirect URL is malformed
Currently, if a return URL doesn't match a page, RedirectToPage throws an InvalidOperationException and we are shown the default error page.

Instead, It would be better if we were redirected to a 404 not found error page.

Use the IExceptionHandlerPathFeature interface to examine the error in the error page model, providing a custom message and correct status code to the response.