Override the client ID for production
We have 2 registered Azure AD applications: SDD Prepare For Advisory Board Non Production, to be used locally, on Dev and on Staging, as well as SDD Prepare For Advisory Board Production for Prod.

We need to override the client ID (specific to the application) in the Prod config,, the rest of the settings are the same across the two apps.