Use authorization header for Cypress tests
We had some problems with the Cypress tests now that we are using Azure SSO.

The solution we came up with in Transfers was to add a new authorization handler, configured to use the authentication header. These handlers are evaluated on an OR basis, so when the correct secret is available in the Cypress tests, we can bypass the SSO implementation.

We set the authorisation header in a global `beforeEach` by intercept in all requests using the `cy.intercept` command.

This includes a fairly big refactor away from maintaining state across multiple tests, which was causing issues with the new authentication mechanism. Cypress suggest not coupling tests together, instead moving shared setup into before blocks.