Add allowed roles to config
We only want users with the "conversions.create" role to have access. Currently, this is everybody in the AZURE PREPARE FOR ADVISORY BOARD LOGIN group which is shared with Transfers.

Copy over the config from Transfers, which includes cookie configuration and a "Access Denied" page override, which the user will see if they do not have the correct role.