Change which dates we fetch from the API
Currently, we are using `OfstedLastInspection` which can refer to either a full or short inspection. However, we want to display the date of the latest short inspection separately if it is later than the latest full inspection.

Instead, use `DateOfLatestSection8Inspection` and `InspectionEndDate` (which always refers to a full inspection).