Remove option to sign out
Currently, we have a link which will sign the user out of all other applications that use their Microsoft account for authentication (single sign-out). After various discussions, we have decided that there is no need for this, and we are happy for it to be removed. 

The service will therefore appear seamless, users will have access when signed in elsewhere (or will be redirected to the sign in page when not), but they will have to sign out somewhere else.