Fix redirect URI on deployed Dev environment
We are seeing an error message when trying to sign into the deployed Dev environment `The redirect URI specified in the request does not match the redirect URIs configured for the application`.

The redirect URI in the error message has the http scheme, which is not supported in Azure. It looks like this has already been fixed in Transfers, hopefully the same will work here.