Suppress view override warnings
We override the MicrosoftIdentity access-denied view in order to provide a custom message.

Unfortunately, we now see the warning `The type 'Areas_MicrosoftIdentity_Pages_Account_AccessDenied' conflicts with the imported type`. It seems as though this is expected and the only choices we have are to ignore it, or supress the warning, so choosing to suppress.